Allow reuse of pagination parameter warping logic from InvocationsProvider.
