+++ client/src/components/providers/InvocationsProvider.js
@@ -1,21 +1,9 @@
 import axios from "axios";
-import { snakeCase } from "snake-case";
+import { cleanPaginationParameters} from "./utils";
 
 export function invocationsProvider(ctx, callback, extraParams) {
     const { apiUrl, ...requestParams } = ctx;
-    const cleanParams = {};
-    Object.entries(requestParams).map(([key, val]) => {
-        if (key === "perPage") {
-            key = "limit";
-        }
-        if (val) {
-            cleanParams[snakeCase(key)] = val;
-        }
-    });
-    if (cleanParams.current_page && cleanParams.limit) {
-        cleanParams.offset = (cleanParams.current_page - 1) * cleanParams.limit;
-        delete cleanParams.current_page;
-    }
+    const cleanParams = cleanPaginationParameters(requestParams);
     const promise = axios.get(apiUrl, { params: { ...cleanParams, ...extraParams } });
 
     // Must return a promise that resolves to an array of items
